adding command to 'vent' the doors
Venting in this context means to open the door ~6 inches up from the floor.  The height should be low for security, but high enough so that a subsequent 'close' command does not bounce the door off the floor into the full up position.  This requires some testing on the door controller in use.  In my case, I have two different door controllers that behave a bit differently on the bounce detect.